feature : 영상 조회수,광고 조회수에 따른 일별,주별,월별,전체 정산 기능 추가

+++ module-batch/src/main/java/com/billing/dto/VideoTop5DurationDTO.java
@@ -8,10 +8,10 @@ import lombok.Setter;
 @Setter
 public class VideoTop5DurationDTO {
 	private String title;
-	private int duration;
+	private long duration;
 
 	@Builder
-	public VideoTop5DurationDTO(String title, int duration) {
+	public VideoTop5DurationDTO(String title, long duration) {
 		this.title = title;
 		this.duration = duration;
 	}
