+++ module-batch/src/main/java/com/billing/dto/VideoTop5ViewDTO.java
@@ -8,10 +8,10 @@ import lombok.Setter;
 @Setter
 public class VideoTop5ViewDTO {
 	private String title;
-	private int viewCount;
+	private long viewCount;
 
 	@Builder
-	public VideoTop5ViewDTO(String title, int viewCount) {
+	public VideoTop5ViewDTO(String title, long viewCount) {
 		this.title = title;
 		this.viewCount = viewCount;
 	}
